continuing to dig down to resolve this openmp failure case when you have many threads simultaneously crossing the export threshold

+++ structure/fof.c
@@ -358,7 +358,7 @@ void fof_find_groups(void)
   Ngblist = (int *) mymalloc("Ngblist", NumPart * sizeof(int));
 
     size_t MyBufferSize = All.BufferSize;
-    All.BunchSize = (int) ((MyBufferSize * 1024 * 1024) / (sizeof(struct data_index) + sizeof(struct data_nodelist) +
+    All.BunchSize = (long) ((MyBufferSize * 1024 * 1024) / (sizeof(struct data_index) + sizeof(struct data_nodelist) +
 					     2 * sizeof(struct fofdata_in)));
     DataIndexTable = (struct data_index *) mymalloc("DataIndexTable", All.BunchSize * sizeof(struct data_index));
     DataNodeList = (struct data_nodelist *) mymalloc("DataNodeList", All.BunchSize * sizeof(struct data_nodelist));
@@ -1442,7 +1442,7 @@ void fof_find_nearest_dmparticle(void)
   Ngblist = (int *) mymalloc("Ngblist", NumPart * sizeof(int));
 
     size_t MyBufferSize = All.BufferSize;
-    All.BunchSize = (int) ((MyBufferSize * 1024 * 1024) / (sizeof(struct data_index) + sizeof(struct data_nodelist) +
+    All.BunchSize = (long) ((MyBufferSize * 1024 * 1024) / (sizeof(struct data_index) + sizeof(struct data_nodelist) +
 					     sizeof(struct fofdata_in) + sizeof(struct fofdata_out) +
 					     sizemax(sizeof(struct fofdata_in), sizeof(struct fofdata_out))));
     DataIndexTable = (struct data_index *) mymalloc("DataIndexTable", All.BunchSize * sizeof(struct data_index));
